+++ src/renderers/raytracing/hittable_list.hpp
@@ -15,7 +15,7 @@ public:
   void clear() { objects.clear(); }
   void add(std::shared_ptr<hittable> object) { objects.push_back(object); }
 
-  virtual bool hit(ray const &r, double t_min, double t_max,
+  virtual bool hit(ray const &r, float t_min, float t_max,
                    hit_record &rec) const override;
 
 public:
