+++ src/renderers/raytracing/sphere.hpp
@@ -1,22 +1,22 @@
 #pragma once
 
 #include "hittable.hpp"
-#include "vec3.hpp"
+#include "../../vec3.hpp"
 
 namespace ren {
 
 class sphere : public hittable {
 public:
   sphere() {}
-  sphere(point3 cen, double r, std::shared_ptr<material> m)
+  sphere(point3 cen, float r, std::shared_ptr<material> m)
       : center(cen), radius(r), mat_ptr(m){};
 
-  virtual bool hit(ray const &r, double t_min, double t_max,
+  virtual bool hit(ray const &r, float t_min, float t_max,
                    hit_record &rec) const override;
 
 public:
   point3 center;
-  double radius;
+  float radius;
   std::shared_ptr<material> mat_ptr;
 };
